+++ src/floating-select/floating-select.component.jsx
@@ -1,46 +1,17 @@
-/* eslint-disable react/no-find-dom-node, prefer-rest-params */
 import React from 'react';
 import Select from 'react-select';
-import ReactDOM from 'react-dom';
-import 'react-select/dist/react-select.css';
+import styles from './styles';
+import FloatingSelectBase from './floating-select-base.component';
 
-import TetherComponent from './tether.component';
-import './floating-select.component.scss';
-
-// Make react-select to tether over the container component instead of stretching it.
-// https://github.com/JedWatson/react-select/issues/810#issuecomment-248546293
-
-export default class FloatingSelect extends Select {
-  constructor(props) {
-    super(props);
-    this.renderOuter = this.renderOuterOverride;
-  }
-
-  componentDidMount() {
-    super.componentDidMount.call(this);
-    this.dropdownFieldNode = ReactDOM.findDOMNode(this);
-  }
-
-  renderOuterOverride() {
-    const menu = super.renderOuter.apply(this, arguments);
-    const options = {
-      attachment: 'top left',
-      targetAttachment: 'bottom left',
-      constraints: [
-        {
-          to: 'window',
-          attachment: 'together',
-        },
-      ],
-    };
-    return (
-      <TetherComponent
-        target={this.dropdownFieldNode}
-        options={options}
-        matchWidth
-      >
-        {React.cloneElement(menu, { style: { position: 'static' } })}
-      </TetherComponent>
-    );
-  }
+export default class FloatingSelect extends FloatingSelectBase {
+  render = () => (
+    <Select
+      {...this.props}
+      components={this.state.components}
+      menuPlacement="auto"
+      onMenuOpen={this.handleOpenMenu}
+      onMenuClose={this.handleCloseMenu}
+      styles={styles(this.props.controlHeight)}
+    />
+  );
 }